Fix audit findings and add delivery-grade export, presets and metering
Fixes:
- Auto EQ + oversampling: the pre-rendered output is now positioned by
  time (rate ratio + interpolation), so it is correct at any factor.
- Multiband limiter: the crossover is prepared eagerly and its latency
  is always reported, so renders are no longer shifted by ~23 ms.
- EQ band edits, fades and band add/remove now trigger the debounced
  chain re-analysis, so playback always matches the export.
- Live playback compensates chain latency: audible-position cursor and
  meters, end-of-song tail flush, and a time-aligned A/B bypass.
- Export sample-rate conversion uses the polyphase windowed-sinc
  resampler (anti-aliased) and re-anchors the true-peak ceiling.
- Leveler measures per-channel K-weighted power (BS.1770); broadband
  limiter true-peak map aligned; 64-bit sample counts in prepare().

Added:
- TPDF dither on 16/24-bit renders and the monitor path.
- Chain presets (JSON), A/B settings slots, parameter undo/redo with a
  byte-bounded history.
- Batch export with per-file analysis, progress and cancellation.
- Metadata preservation (ID3, LIST-INFO, bext) on same-container exports.
- Selection looping; true-peak meter; phase correlation, M/S levels and
  a goniometer.

Changed:
- One-pass cumulative chain analysis with a post-EQ prefix cache; the
  output meters reuse the same render.
- Rate-independent linear-phase kernels (EQ + crossover) and ADAA on
  all clip curves (updated vendored dspark jar).

+++ src/main/java/com/quickmaster/processing/AudioProcessor.java
@@ -18,7 +18,7 @@ package com.quickmaster.processing;
  * by passing the entire file as a single buffer.
  * <p>
  * <b>Lifecycle.</b> Before the first buffer is processed,
- * {@link #prepare(int, int)} must be called to inform the
+ * {@link #prepare(int, long)} must be called to inform the
  * processor of the audio sample rate and total sample count.
  * This pattern, borrowed from professional audio plugin APIs,
  * separates one-off setup work (such as recomputing filter
@@ -85,10 +85,12 @@ public interface AudioProcessor
      * @param totalSamples  total number of float samples in the
      *                      whole audio, summed across channels
      *                      (e.g. 1 second of 44.1 kHz stereo =
-     *                      88200). Use 0 if unknown or not
-     *                      applicable to the implementation.
+     *                      88200). Declared {@code long} so an
+     *                      oversampled total cannot overflow on
+     *                      long, high-rate files. Use 0 if unknown
+     *                      or not applicable to the implementation.
      */
-    void prepare(int sampleRate, int totalSamples);
+    void prepare(int sampleRate, long totalSamples);
 
     /**
      * Optional whole-audio pre-scan. The pipeline calls this
